+++ src/schedule.jl
@@ -8,9 +8,9 @@
 
 """
 ```
-event!([clk::CL], ex::Action, t::Number; scope::Module=Main, cycle::Number=0.0,
+event!([clk::CL], ex::Action, t::Number; cycle::Number=0.0,
         cid::Int=clk.id, spawn=false, sync::Bool=false) where {CL<:AbstractClock}
-event!([clk::CL], ex::Action, T::Timing, t::Number; scope::Module=Main,
+event!([clk::CL], ex::Action, T::Timing, t::Number;
         cid::Int=clk.id, spawn=false, sync::Bool=false) where {CL<:AbstractClock}
 ```
 Schedule an event for a given simulation time.
@@ -22,7 +22,6 @@ Schedule an event for a given simulation time.
 - `t::Real` or `t::Time`: simulation time, if t < clk.time set t = clk.time,
 
 # Keyword arguments
-- `scope::Module=Main`: scope for expressions to be evaluated in,
 - `cycle::Float64=0.0`: repeat cycle time for an event,
 - `cid::Int=clk.id`: if cid ≠ clk.id, assign the event to the parallel clock
     with id == cid. This overrides `spawn`,
@@ -60,7 +59,7 @@ julia> event!(fun(myfunc, 5, 6), after, 1hr)
 
 ```
 """
-function event!(clk::CL, ex::Action, t::Number; scope::Module=Main, cycle::Number=0.0,
+function event!(clk::CL, ex::Action, t::Number; cycle::Number=0.0,
                 cid::Int=clk.id, spawn::Bool=false, sync::Bool=false) where {CL<:AbstractClock}
     t = tadjust(clk, t)
     cycle = tadjust(clk, cycle)
@@ -69,17 +68,17 @@ function event!(clk::CL, ex::Action, t::Number; scope::Module=Main, cycle::Numbe
     if cid == clk.id && spawn  # evaluate spawn only if cid == clk.id
         cid = spawnid(clk)
     end
-    assign(clk, DiscreteEvent(ex, scope, t, cycle), cid)
+    assign(clk, DiscreteEvent(ex, t, cycle), cid)
 end
-function event!(clk::CL, ex::Action, T::Timing, t::Number; scope::Module=Main,
+function event!(clk::CL, ex::Action, T::Timing, t::Number;
                 cid::Int=clk.id, spawn::Bool=false, sync::Bool=false) where {CL<:AbstractClock}
     t = tadjust(clk, t)
     if T == after
-        event!(clk, ex, t+clk.time, scope=scope, cid=cid, spawn=spawn, sync=sync)
+        event!(clk, ex, t+clk.time, cid=cid, spawn=spawn, sync=sync)
     elseif T == every
-        event!(clk, ex, clk.time, scope=scope, cycle=t, cid=cid, spawn=spawn, sync=sync)
+        event!(clk, ex, clk.time, cycle=t, cid=cid, spawn=spawn, sync=sync)
     else
-        event!(clk, ex, t, scope=scope, cid=cid, spawn=spawn, sync=sync)
+        event!(clk, ex, t, cid=cid, spawn=spawn, sync=sync)
     end
 end
 event!(ex::Action, t::Number; kw...) = event!(𝐶, ex, t; kw...)
@@ -88,7 +87,7 @@ event!(ex::Action, T::Timing, t::Number; kw...) = event!(𝐶, ex, T, t; kw...)
 
 """
 ```
-event!([clk::T], ex::Action, cond::Action; scope::Module=Main,
+event!([clk::T], ex::Action, cond::Action;
         cid::Int=clk.id, spawn=false) where {T<:AbstractClock}
 ```
 Schedule a conditional event.
@@ -105,7 +104,6 @@ sampling rate is setup depending on the scale of the remaining simulation time
 - `cond::Union{SimExpr, Tuple{SimExpr}}`: a condition is an expression or function
     or a tuple of them. It is true only if all expressions or `fun`s
     therein return true,
-- `scope::Module=Main`: scope for the expressions to be evaluated
 - `cid::Int=clk.id`: if cid ≠ clk.id, assign the event to the parallel clock
     with id == cid. This overrides `spawn`,
 - `spawn::Bool=false`: if true, spawn the event at other available threads.
@@ -133,15 +131,15 @@ Clock thread 1 (+ 0 ac): state=Simulate.Idle(), t=10.0 , Δt=0.0 , prc:0
   scheduled ev:0, cev:0, sampl:0
 ```
 """
-function event!(clk::T, ex::Action, cond::Action; scope::Module=Main,
+function event!(clk::T, ex::Action, cond::Action;
                 cid::Int=clk.id, spawn=false) where {T<:AbstractClock}
-    if busy(clk) && all(evExec(cond))   # all conditions met
-        evExec(ex)                      # execute immediately
+    if busy(clk) && all(evaluate(cond))   # all conditions met
+        evaluate(ex)                      # execute immediately
     else
         if cid == clk.id && spawn  # evaluate spawn only if cid == clk.id
             cid = spawnid(clk)
         end
-        assign(clk, DiscreteCond(cond, ex, scope), cid)
+        assign(clk, DiscreteCond(cond, ex), cid)
     end
     return nothing
 end
@@ -149,8 +147,7 @@ event!( ex::Action, cond::Action; kw...) = event!(𝐶, ex, cond; kw...)
 
 """
 ```
-periodic!([clk::Clock], ex::T, Δt::Number=clk.Δt;
-        scope::Module=Main, spawn=false) where {T<:Action}
+periodic!([clk::Clock], ex::T, Δt::Number=clk.Δt; spawn=false) where {T<:Action}
 periodic!(ac::ActiveClock, ex::T, Δt::Number=ac.clock.Δt; kw...) where {T<:Action}
 ```
 Register a function or expression for periodic execution at the clock`s sample rate.
@@ -160,12 +157,11 @@ Register a function or expression for periodic execution at the clock`s sample r
 - `ex<:Action`: an expression or function or a tuple of them,
 - `Δt::Number=clk.Δt`: set the clock's sampling rate, if no Δt is given, it takes
     the current sampling rate, if that is 0, it calculates one,
-- `scope::Module=Main`: optional, an evaluation scope for a given expression.
 """
 function periodic!(clk::Clock, ex::T, Δt::Number=clk.Δt;
-                 scope::Module=Main, spawn=false) where {T<:Action}
+                   spawn=false) where {T<:Action}
     clk.Δt = Δt == 0 ? scale(clk.end_time - clk.time)/100 : Δt
-    assign(clk, Sample(ex, scope), spawn ? spawnid(clk) : 0)
+    assign(clk, Sample(ex), spawn ? spawnid(clk) : 0)
 end
 periodic!(ex::T, Δt::Number=𝐶.Δt; kw...) where {T<:Action} = periodic!(𝐶, ex, Δt; kw...)
 periodic!(ac::ActiveClock, ex::T, Δt::Number=ac.clock.Δt; kw...) where {T<:Action} = periodic!(ac.clock, ex, Δt; kw...)
@@ -216,7 +212,7 @@ Assign an event to a clock.
 
 There are several ways to do it:
 1. assign it directly to a clock or an active clock or
-2. assign it via a clock via a clock to another one, if given a different id.
+2. assign it via a channel to another one, if given a different id.
     In this case the event is sent over the channel to the target clock.
     - The master clock (id=0) can directly send to an active clock.
     - An active clock can send directly to master.
@@ -224,7 +220,7 @@ There are several ways to do it:
 """
 function assign(c::S, ev::T, id::Int=c.id) where {S<:AbstractClock, T<:AbstractEvent}
     if id == c.id
-        register!(c, ev)       # 1: register it yourself
+        register!(c, ev)       # 1: register it to yourself
     else
         register(c, ev, id)    # 2: register it to another clock
     end
@@ -264,9 +260,9 @@ Register an event to another clock via a channel.
 
 # Arguments
 - `c::Clock`: a master clock can forward events to active clocks,
-- `ac::ActiveClock`: active clocks can forward events only through master, he then
-    does the distribution for them,
-- `ev::AbstractEvent`: the event to register,
+- `ac::ActiveClock`: active clocks can forward events only through master,
+    he then does the distribution for them,
+- `ev<:AbstractEvent`: the event to register,
 - `id::Int`: the id of the clock it should get registered to.
 """
 function register(c::Clock, ev::T, id::Int) where {T<:AbstractEvent}