+++ test/test_threads1.jl
@@ -1,4 +1,6 @@
-println("... testing multithreading  1 ...")
+const sleeptime = 0.3
+
+println("... testing multithreading  1, (sleeptime=$sleeptime) ...")
 println("number of available threads: ", nthreads())
 
 clk = PClock()
@@ -24,16 +26,16 @@ err = diag(clk, 1)
 
 println("... testing channel and active clock ...")
 a = 0
-ev = Simulate.DiscreteEvent(fun(()->global a+=1),Main,1.0,0.0)
+ev = Simulate.DiscreteEvent(fun(()->global a+=1),1.0,0.0)
 put!(clk.ac[1].forth, Simulate.Register(ev))
 sleep(0.1)
 @test length(c1.clock.sc.events) == 1
-cev = Simulate.DiscreteCond(fun(≥, fun(tau, c1), 5), fun(()->global a+=1), Main)
+cev = Simulate.DiscreteCond(fun(≥, fun(tau, c1), 5), fun(()->global a+=1))
 put!(clk.ac[1].forth, Simulate.Register(cev))
 sleep(0.1)
 @test length(c1.clock.sc.cevents) == 1
 b = 0
-sp = Simulate.Sample(fun(()->global b+=1), Main)
+sp = Simulate.Sample(fun(()->global b+=1))
 put!(clk.ac[1].forth, Simulate.Register(sp))
 sleep(0.1)
 @test length(c1.clock.sc.samples) == 1
@@ -46,27 +48,31 @@ sleep(0.5)
 @test a == 2
 @test b == 1000
 
-println("... testing register(!) five cases ... ")
-reset!(c1.clock)
+println("... parallel clock 1 reset ...")
+put!(clk.ac[1].forth, Simulate.Reset(true))
+@test take!(clk.ac[1].back).x == 1
 @test c1.clock.time == 0
-ev1 = Simulate.DiscreteEvent(fun(()->global a+=1),Main,1.0,0.0)
-ev2 = Simulate.DiscreteEvent(fun(()->global a+=1),Main,2.0,0.0)
+@test c1.clock.scount == 0
+
+println("... testing register(!) five cases ... ")
+ev1 = Simulate.DiscreteEvent(fun(()->global a+=1),1.0,0.0)
+ev2 = Simulate.DiscreteEvent(fun(()->global a+=1),2.0,0.0)
 Simulate.register(clk, ev1, 0)              # 1. register ev1 to clk
 @test Simulate.nextevent(clk) == ev1
 Simulate.register(clk, ev1, 1)              # 2. register ev1 to 1st parallel clock
-sleep(0.1)
+sleep(sleeptime)
 @test Simulate.nextevent(c1.clock) == ev1
 Simulate.register(c1, ev2, 1)               # 3. register ev2 directly to 1st parallel clock
-sleep(0.1)
+sleep(sleeptime)
 @test length(c1.clock.sc.events) == 2
 Simulate.register(c1, ev2, 0)               # 4. register ev2 back to master
-sleep(0.1)
+sleep(sleeptime)
 @test length(clk.sc.events) == 2
 
 if nthreads() > 2                           # This fails on CI (only 2 threads)
     Simulate.register(c1, ev2, 2)           # 5. register ev2 to another parallel clock
     c2 = pclock(clk, 2)
-    sleep(0.1)
+    sleep(sleeptime)
     @test Simulate.nextevent(c2.clock) == ev2
 end